.README 파일추가 .ContextPath 변경 및 관련코드 수정 .index.html추가

+++ src/main/java/com/yorath/booksearch/dto/NaverBookSearchResponseDto.java
@@ -31,8 +31,8 @@ public class NaverBookSearchResponseDto {
         private String isbn;
         private String pubdate;
         private String publisher;
-        private int discount;
-        private int price;
+        private double discount;
+        private double price;
         private String author;
         private String image;
         private String link;
